don't call the Garbage Collector inside the lock
If there a a lot of documents (like ~2000 in Tiger2) the old behavior will result in a very costly repeated garbage collection at the beginning before any document was converted. By putting the the call behind the lock the hint to the GC is still several times, but the GC is smart enough to only execute garbage collection a few times.