fixed #92
the bug was caused by the maven filter plugin, which is configured to
replace all maven variables with their values. To fix that I excluded
the entire file (dependencies.md) from the replacement mechanism. This
seems to be a bit exaggerated. Better would be a possibility to mark the
use of the variable itself not to be resolved. If any one knows such a
mechanism, please tell me.