Use a thread-safe hash set implementation for controller list and also make sure there is no reace condition when creating a new instance in getter function.
The original getter function checked for null first and then synchronized an only synchronized the access when a new instance was created. There was the possibility of race condition that an existing instance was overwritten by a new one when getControllerList() was called at the same time and the latter call would overwrite the original instance. This is solved by making the member variable final and creating the hash set on construction time of the ModuleControllerImpl.